review: rename info module to config
